refactor: move document property builders into separate files
