+++ catalyst_voices/apps/voices/lib/widgets/document_builder/value/single_grouped_tag_selector_widget.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/widgets/document_builder/document_error_text.dart';
+import 'package:catalyst_voices/widgets/document_builder/value/document_error_text.dart';
 import 'package:catalyst_voices/widgets/dropdown/voices_dropdown.dart';
 import 'package:catalyst_voices/widgets/form/voices_form_field.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
