+++ catalyst_voices/apps/voices/lib/widgets/document_builder/value/yes_no_choice_widget.dart
@@ -1,4 +1,4 @@
-import 'package:catalyst_voices/widgets/document_builder/document_error_text.dart';
+import 'package:catalyst_voices/widgets/document_builder/value/document_error_text.dart';
 import 'package:catalyst_voices/widgets/form/voices_form_field.dart';
 import 'package:catalyst_voices/widgets/widgets.dart';
 import 'package:catalyst_voices_localization/catalyst_voices_localization.dart';
@@ -27,7 +27,9 @@ class YesNoChoiceWidget extends StatefulWidget {
 
 class _YesNoChoiceWidgetState extends State<YesNoChoiceWidget> {
   bool? get _value => widget.property.value ?? widget.schema.defaultValue;
+
   String get _title => widget.schema.title;
+
   bool get _isRequired => widget.schema.isRequired;
 
   @override
