+++ catalyst_voices/apps/voices/lib/widgets/widgets.dart
@@ -29,10 +29,12 @@ export 'containers/space_side_panel.dart';
 export 'containers/workspace_text_tile_container.dart';
 export 'document/document_tag_chip.dart';
 export 'document/document_version_selector.dart';
+export 'document_builder/document_property_builder.dart';
 export 'drawer/voices_drawer.dart';
 export 'drawer/voices_drawer_space_chooser.dart';
 export 'footers/links_page_footer.dart';
 export 'footers/standard_links_page_footer.dart';
+export 'form/voices_form_field.dart';
 export 'headers/brand_header.dart';
 export 'headers/section_header.dart';
 export 'headers/segment_header.dart';
